added helium_* calling from erlang

+++ c_src/include/helium.h
@@ -11,7 +11,7 @@ __BEGIN_DECLS
 Currently, program will quit when initialized failed
  
 */
-int8_t helium_init();
+int helium_init();
 
 /*
   return: 
@@ -19,14 +19,14 @@ int8_t helium_init();
     0: not initialized,
 */
 
-int8_t helium_deinit();
+int helium_deinit();
 
 /*
   generate key
 */
 
-int8_t gen_ecdsa_keypair();
-int8_t ecdsa_sign(void* inbuf, size_t inbuf_len, void* outbuf, size_t* outbuf_len);
+int gen_ecdsa_keypair();
+int ecdsa_sign(void* inbuf, size_t inbuf_len, void* outbuf, size_t* outbuf_len);
 __END_DECLS
 
 #endif /* HELIUM_H */
